A bit of a cleanup since we're now compiling primarily for Java 5: clone() returns objects of called type, so no casting is necessary anymore

+++ src/org/mozilla/javascript/NativeArray.java
@@ -1576,6 +1576,5 @@ public class NativeArray extends IdScriptableObject
      * The factor to grow <code>dense</code> by.
      */
     private static final double GROW_FACTOR = 1.5;
-    private static final int MAX_PRE_GROW_SIZE =
-            (int)((double)Integer.MAX_VALUE / GROW_FACTOR);
+    private static final int MAX_PRE_GROW_SIZE = (int)(Integer.MAX_VALUE / GROW_FACTOR);
 }
